+++ src/auxiliary/setup.cpp
@@ -16,9 +16,17 @@ void setDeltaT(Foam::Time& ofRunTime, RunTime& nfRunTime, Foam::scalar coNum)
     Foam::scalar deltaTFact = Foam::min(Foam::min(maxDeltaTFact, 1.0 + 0.1 * maxDeltaTFact), 1.2);
 
     ofRunTime.setDeltaT(Foam::min(deltaTFact * ofRunTime.deltaTValue(), nfRunTime.maxDeltaT));
-    Foam::Info << "deltaT = " << ofRunTime.deltaTValue() << Foam::endl;
+    NeoN::Logging::info("deltaT = {}", ofRunTime.deltaTValue());
+}
 
+void syncRunTimes(Foam::Time& ofRunTime, RunTime& nfRunTime, Foam::scalar coNum)
+{
+    if (nfRunTime.adjustTimeStep)
+    {
+        setDeltaT(ofRunTime, nfRunTime, coNum);
+    }
     nfRunTime.dt = ofRunTime.deltaTValue();
+    nfRunTime.t = ofRunTime.time().value();
 }
 
 
@@ -32,9 +40,8 @@ std::unique_ptr<MeshAdapter> createMesh(const NeoN::Executor& exec, const Foam::
 std::unique_ptr<Foam::fvMesh> createMesh(const Foam::Time& runTime)
 {
     std::unique_ptr<Foam::fvMesh> meshPtr;
-    Foam::Info << "Create mesh";
     Foam::word regionName(Foam::polyMesh::defaultRegion);
-    Foam::Info << " for time = " << runTime.timeName() << Foam::nl;
+    NeoN::Logging::info("Create mesh for time = {}", runTime.timeOutputValue());
 
     meshPtr.reset(new Foam::fvMesh(
         Foam::IOobject(regionName, runTime.timeName(), runTime, Foam::IOobject::MUST_READ),
@@ -48,22 +55,19 @@ std::unique_ptr<Foam::fvMesh> createMesh(const Foam::Time& runTime)
 /* @brief create a NeoN executor from a name
  * @return the Neon::Executor
  */
-NeoN::Executor createExecutor(const Foam::word& execName)
+NeoN::Executor createExecutor(const std::string execName)
 {
-    Foam::Info << "Creating Executor: " << execName << Foam::endl;
+    NeoN::Logging::info("Creating Executor {}", execName);
     if (execName == "Serial")
     {
-        Foam::Info << "Serial Executor" << Foam::endl;
         return NeoN::SerialExecutor();
     }
     if (execName == "CPU")
     {
-        Foam::Info << "CPU Executor" << Foam::endl;
         return NeoN::CPUExecutor();
     }
     if (execName == "GPU")
     {
-        Foam::Info << "GPU Executor" << Foam::endl;
         return NeoN::GPUExecutor();
     }
     Foam::FatalError << "unknown Executor: " << execName << Foam::nl
@@ -75,7 +79,7 @@ NeoN::Executor createExecutor(const Foam::word& execName)
 
 NeoN::Executor createExecutor(const Foam::dictionary& dict)
 {
-    auto execName = dict.get<Foam::word>("executor");
+    auto execName = std::string(dict.get<Foam::word>("executor"));
     return createExecutor(execName);
 }
 
@@ -87,8 +91,7 @@ NeoFOAM::RunTime createAdapterRunTime(const Foam::Time& in)
 
 RunTime createAdapterRunTime(const Foam::Time& in, const NeoN::Executor exec)
 {
-    std::cout << __FILE__ << ":"
-              << "Creating NeoFOAM runTime\n";
+    NeoN::Logging::info("Creating NeoFOAM runTime");
     std::unique_ptr<MeshAdapter> meshPtr = createMesh(exec, in);
     MeshAdapter& mesh = *meshPtr;